pyproject.toml
[project]
name = "gcp-super-dev"
version = "0.1.0"
description = "This project generate from 0 to 10 commits to a Github repository"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Geoffrey Louasse"}]
dependencies = [
    "loguru",
    "faker",
    "python-dotenv"
]
